Another bunch of changes including: Massive potencial field improvement and coefficent adjustment bugfixes performance improvement by reducing calculation of formation parameters point capture behavior

+++ enhanced.nim
@@ -6,7 +6,7 @@ const maxsize* = (1024 div gridsize)
 type
   Group* = uint8
   VehicleId* = uint16
-  FacilityId* = uint16
+  FacilityId* = uint8
   EVehicle* = tuple
     id: int64
     sid: VehicleId
